+++ src/js/layout/components/MenuDrawer.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router';
+import { withRouter } from 'react-router-dom';
 import Drawer from 'material-ui/Drawer';
 import Divider from 'material-ui/Divider';
 import List, { ListItem, ListItemIcon, ListItemText } from 'material-ui/List';
@@ -17,13 +17,18 @@ type Props = {
   isMenuCurrentlyOpened: boolean,
   shouldMenuLeftOpened: boolean,
   onToggleMenu: Function,
-  router: Object,
+  history: Object,
   classes: Object,
 };
 
-const MenuDrawer = ({
-  isMenuCurrentlyOpened, shouldMenuLeftOpened, onToggleMenu, router, classes,
-}: Props) => {
+const MenuDrawer = ( // eslint-disable-line function-paren-newline
+  {
+    isMenuCurrentlyOpened,
+    shouldMenuLeftOpened,
+    onToggleMenu,
+    history,
+    classes,
+  }: Props) => {
   /**
    * When changing route, determine if there's a need to hide the menu especially when
    * user uses a small viewing device.
@@ -31,7 +36,7 @@ const MenuDrawer = ({
    * @param path    Path to switch to
    */
   const changeRoute: Function = (path: string) => () => {
-    router.push(path);
+    history.push(path);
 
     if (!shouldMenuLeftOpened) {
       onToggleMenu();
@@ -57,14 +62,14 @@ const MenuDrawer = ({
           <ListItemText primary="Home" />
         </ListItem>
 
-        <ListItem button onClick={changeRoute('chuck-norris')}>
+        <ListItem button onClick={changeRoute('/chuck-norris')}>
           <ListItemIcon>
             <MoodIcon />
           </ListItemIcon>
           <ListItemText primary="Chuck Norris" />
         </ListItem>
 
-        <ListItem button onClick={changeRoute('todo-manager')}>
+        <ListItem button onClick={changeRoute('/todo-manager')}>
           <ListItemIcon>
             <AssignmentIcon />
           </ListItemIcon>